# perf(uuid): make `uuidToBytes()` up to 2.5x faster